requirements/dev.txt
-r base.txt

bpython==0.16
Werkzeug==0.12.1
django-debug-toolbar==1.7
